:warning: Implemented JWT authentication
When the user logs in, a JWT is sent and saved as a session cookie
along with a CSRF token. Any POST, PUT, PATCH, or DELETE request
sends this CSRF token to the server along with the JWT obtained
from logging in.

Logging out now also makes a request to the server's /logout route.
This is necessary because it unsets the user's JWT.